refactor(yaml): prepare for serialization implementation
