+++ facet-yaml/src/serialize/error.rs
@@ -1,0 +1,1 @@
+//! Errors from parsing into YAML documents.
